requirements.txt
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
comtypes==1.4.16
cryptography==46.0.7
distro==1.9.0
docopt==0.6.2
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.72.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.14.0
Js2Py==0.74
openai==2.31.0
packaging==26.0
pipwin==0.5.2
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
pyjsparser==2.7.1
pyparsing==3.3.2
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.2.2
pyttsx3==2.99
pywin32==311
requests==2.33.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SpeechRecognition==3.16.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
